+++ src/utils/constantUtils.ts
@@ -2,4 +2,4 @@
 export const DEFAULT_THROTTLE_TIMEOUT = 500;
 
 /** The width of a thumbnail element. */
-export const THUMBNAIL_WIDTH = 250;
+export const THUMBNAIL_WIDTH = 240;