+++ src/webviewer/getThumbnail.ts
@@ -1,10 +1,11 @@
 import { Futurable } from '../data/futurable';
+import { THUMBNAIL_WIDTH } from '../utils/constantUtils';
 
 async function getThumbnail(documentObj: Futurable<CoreControls.Document>) {
   const fetchedDocument = await documentObj;
   const canvas: HTMLCanvasElement = await new Promise(resolve => {
     const pageWidth = fetchedDocument.getPageInfo(0).width;
-    const zoom = 250 / pageWidth;
+    const zoom = THUMBNAIL_WIDTH / pageWidth;
     fetchedDocument.loadCanvasAsync({ pageIndex: 0, drawComplete: resolve, zoom });
     fetchedDocument.loadThumbnailAsync(0, resolve);
   });